Shifting to Components On More Details of Full App
